final commit

+++ todo_app/app.py
@@ -3,16 +3,17 @@ from todo_app.data.mongoDB_items import MongoDBTasks
 from todo_app.viewmodel import ViewModel
 from todo_app.user import User
 from todo_app.flask_config import Config
-from flask_login import LoginManager, login_required, login_user
+from flask_login import LoginManager, login_required, login_user, current_user
 import os, requests
+from functools import wraps
 
 
 def create_app():
     app = Flask(__name__)
     mongodb_tasks = MongoDBTasks()
     app.config.from_object(Config())
-
     login_manager = LoginManager()
+
     
     @login_manager.unauthorized_handler
     def unauthenticated():
@@ -47,21 +48,30 @@ def create_app():
             "Authorization": f"Bearer {access_token}"
         }
         id = (requests.get(user_info_url, headers=user_info_headers)).json()['id']
-        print(id)
+        response = requests.get(user_info_url, headers=user_info_headers).json()
+        print(response)
 
+        # Step 3. Create an instance of User using 'id' from Step 2 and login that user.
         user = User(user_id=id)
-        print(user)
         login_user(user)
-
         return redirect(url_for('index'))
 
+    # def authorise_user(f):
+    #     @wraps(f)
+    #     def decorator(*args, **kwargs):
+    #         if current_user.role == 'admin':
+    #             return f(current_user, *args, **kwargs)
+    #     return decorator
 
     @app.route('/')
-    @login_required 
+    @login_required
     def index():
         tasks = mongodb_tasks.get_all_tasks()
         task_view_model = ViewModel(tasks)
-        return render_template('index.html', view_model = task_view_model)
+        if current_user.role == 'admin':
+            return render_template('index.html', view_model = task_view_model)
+        else:
+            return render_template('index_with_hidden_buttons.html', view_model = task_view_model)
 
     @app.route("/add", methods=['Post'])
     @login_required 
@@ -69,38 +79,56 @@ def create_app():
         title = request.form['title']
         desc = request.form['desc']
         dueby = request.form['dueby']
-        mongodb_tasks.add_task(title, desc, due=dueby)
-        return redirect(url_for('index'))
+        if current_user.role == 'admin':
+            mongodb_tasks.add_task(title, desc, due=dueby)
+            return redirect(url_for('index'))
+        else:
+            return render_template('unauthorised.html')
 
     @app.route("/done/<id>")
     @login_required 
     def done(id):
-        mongodb_tasks.mark_as_done(id)
-        return redirect(url_for('index'))
+        if current_user.role == 'admin':
+            mongodb_tasks.mark_as_done(id)
+            return redirect(url_for('index'))
+        else:
+            return render_template('unauthorised.html')
 
     @app.route("/in_progress/<id>")
     @login_required 
     def in_progress(id):
-        mongodb_tasks.mark_as_in_progress(id)
-        return redirect(url_for('index'))
+        if current_user.role == 'admin':
+            mongodb_tasks.mark_as_in_progress(id)
+            return redirect(url_for('index'))
+        else:
+            return render_template('unauthorised.html')
 
     @app.route("/not_started/<id>")
     @login_required 
     def not_started(id):
-        mongodb_tasks.mark_as_not_started(id)
-        return redirect(url_for('index'))
+        if current_user.role == 'admin':
+            mongodb_tasks.mark_as_not_started(id)
+            return redirect(url_for('index'))
+        else:
+            return render_template('unauthorised.html')
 
     @app.route("/delete/<id>")
     @login_required 
     def delete(id):
-        mongodb_tasks.delete_task(id)
-        return redirect(url_for('index'))
+        if current_user.role == 'admin':
+            mongodb_tasks.delete_task(id)
+            return redirect(url_for('index'))
+        else:
+            return render_template('unauthorised.html')
 
     @app.route("/update/<id>")
     @login_required 
     def show_update_page(id):
-        task = mongodb_tasks.get_task(id)
-        return render_template('update.html', id = id, task = task)
+        if current_user.role == 'admin':
+            task = mongodb_tasks.get_task(id)
+            return render_template('update.html', id = id, task = task)
+        else:
+            return render_template('unauthorised.html')
         
     @app.route("/update/<id>/submit", methods = ['POST'])
     @login_required 
@@ -108,15 +136,21 @@ def create_app():
         new_title = request.form['title']
         new_desc = request.form['desc']
         new_due = request.form['due']
-        mongodb_tasks.update_task(id, new_title, new_desc, new_due)
-        return index()
+        if current_user.role == 'admin':
+            mongodb_tasks.update_task(id, new_title, new_desc, new_due)
+            return index()
+        else:
+            return render_template('unauthorised.html')
 
     @app.route('/index_with_all_completed')
     @login_required 
     def index_with_all_completed():
         tasks = mongodb_tasks.get_all_tasks()
         task_view_model = ViewModel(tasks)
-        return render_template('index_with_all_completed.html', view_model = task_view_model)
+        if current_user.role == 'admin':
+            return render_template('index_with_all_completed.html', view_model = task_view_model)
+        else:
+            return render_template('index_with_all_completed_with_hidden_buttons.html', view_model = task_view_model)
 
     return app
